chore: consolidate logic for getUsageDataMetric function
